Fix viewer & move double implementation
